[refactor]: Remove java/lotto/utils/*
