+++ src/main/java/lotto/view/Shuffler.java
@@ -1,4 +1,4 @@
-package lotto.utils;
+package lotto.view;
 
 import java.util.Collections;
 import java.util.LinkedList;
